Backend/requirements.txt
django==4.1.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.5.0
mysqlclient